requirements.txt
docker
flask
flask-sqlalchemy
flask-fixtures
Flask-Mail
pyyaml
python-gitlab
PyGithub
slackclient==1.3.2
websocket-client<0.55
cffi
pytest
